requirements.txt
numpy~=2.3.4
torch~=2.9.0
pillow~=12.0.0
flwr~=1.22.0
pandas~=2.3.3
matplotlib~=3.10.7